ft(retrofit instance): retrofit instance
-Create Retrofit instance

[Starts #155890684]
