+++ .idea/modules.xml
@@ -2,8 +2,8 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/JavaDevelopers.iml" filepath="$PROJECT_DIR$/JavaDevelopers.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
+      <module fileurl="file://$PROJECT_DIR$/java-developers.iml" filepath="$PROJECT_DIR$/java-developers.iml" />
     </modules>
   </component>
 </project>
