+++ app/src/main/java/org/andela/app/javadevelopers/model/GithubUsersResponse.java
@@ -1,0 +1,12 @@
+package org.andela.app.javadevelopers.model;
+
+import java.util.List;
+
+public class GithubUsersResponse {
+
+    List<GithubUsers> items;
+
+    public GithubUsersResponse(List<GithubUsers> items) {
+        this.items = items;
+    }
+}
