+++ app/src/main/java/org/andela/app/javadevelopers/service/GithubService.java
@@ -1,0 +1,16 @@
+package org.andela.app.javadevelopers.service;
+
+import retrofit2.Retrofit;
+import retrofit2.converter.gson.GsonConverterFactory;
+
+public class GithubService {
+
+        String BASE_URL = "https://api.github.com/";
+
+            Retrofit retrofit = new Retrofit
+                    .Builder()
+                    .baseUrl(BASE_URL)
+                    .addConverterFactory(GsonConverterFactory.create())
+                    .build();
+
+}
