+++ app/src/main/java/org/andela/app/javadevelopers/view/GithubUsersView.java
@@ -1,0 +1,10 @@
+package org.andela.app.javadevelopers.view;
+
+import org.andela.app.javadevelopers.model.GithubUsers;
+
+import java.util.List;
+
+public interface GithubUsersView {
+
+    void githubUsersReady(List<GithubUsers> githubUsersList);
+}
